requirements.txt
fastapi==0.115.12
fastapi-cli==0.0.7
line-bot-sdk==3.17.1
ollama==0.6.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
SQLAlchemy==2.0.40
sqlmodel==0.0.24
uvicorn==0.34.2